add authorization to url related actions

+++ src/routers/url.ts
@@ -1,7 +1,10 @@
 import { Router } from "express";
 import { nanoid } from "nanoid";
+import { Container } from "typedi";
 import * as yup from "yup";
 
+import { ShortURLService } from "../db/short-url/service";
+import { withAuth } from "../middlewares/withAuth";
 import { CodedError } from "../utils/errors/CodedError";
 
 export const urlRouter = Router();
@@ -14,14 +17,17 @@ const schema = yup.object().shape({
   url: yup.string().trim().url().required()
 });
 
-urlRouter.get("/", async (req, res) => {
+urlRouter.get("/", withAuth, async (req, res) => {
   const urls = await req.shortURLsRepository.find({
+    where: {
+      user: req.loggedInUser
+    },
     order: { lastUpdated: "DESC" }
   });
   res.send(urls);
 });
 
-urlRouter.post("/", async (req, res, next) => {
+urlRouter.post("/", withAuth, async (req, res, next) => {
   try {
     let { alias, url } = req.body;
     await schema.validate({
@@ -43,9 +49,10 @@ urlRouter.post("/", async (req, res, next) => {
       } while (await req.shortURLsRepository.findOne({ alias }));
     }
 
-    const created = await req.shortURLsRepository.save({
+    const created = await req.shortURLsRepository.insert({
       url,
-      alias
+      alias,
+      user: req.loggedInUser
     });
 
     res.status(200).json(created);
@@ -54,9 +61,13 @@ urlRouter.post("/", async (req, res, next) => {
   }
 });
 
-urlRouter.patch("/", async (req, res, next) => {
+urlRouter.patch("/", withAuth, async (req, res, next) => {
   try {
     const { url, alias } = req.body;
+    const shortURLsService = Container.get(ShortURLService);
+    if (!(await shortURLsService.urlBelongsToUser({ alias }, req.loggedInUser))) {
+      throw new CodedError("User is not allowed to edit provided url", 401);
+    }
     await schema.validate({ url });
     await req.shortURLsRepository.update({ alias }, { url });
     res.json({ url, alias });
@@ -65,9 +76,13 @@ urlRouter.patch("/", async (req, res, next) => {
   }
 });
 
-urlRouter.delete("/:id", async (req, res, next) => {
+urlRouter.delete("/:id", withAuth, async (req, res, next) => {
   try {
     const { id } = req.params;
+    const shortURLsService = Container.get(ShortURLService);
+    if (!(await shortURLsService.urlBelongsToUser({ id }, req.loggedInUser))) {
+      throw new CodedError("User is not allowed to delete provided url", 401);
+    }
     await req.shortURLsRepository.delete(id);
     res.sendStatus(200);
   } catch (error) {
